- ITK 5 with    ITK_LEGACY_REMOVE ON  - C++ 17  - remove many warnings, especialy MSVC with sprintf, strcpy, etc  - avoid warning DTK on MSVC about unknown pragma  - increase Windows Git-Bash detection
Next step:
 - improve and factorize correction about sprintf, strcpy

+++ src/layers/medCore/data/medDiffusionGradientReader.cpp
@@ -141,7 +141,7 @@ void medDiffusionGradientReader::readBVecFile(QString gradientFileName)
         while (!iss.eof());
     }
 
-    unsigned int numDirections = trVecs[0].size();
+    unsigned int numDirections = static_cast<unsigned int>(trVecs[0].size());
 
     d->gradients.resize(numDirections);
     for (unsigned int i = 0;i < numDirections;++i)
